code: move macro to setup
